chore: Move logger to utils
